+++ rust/patchable/src/main.rs
@@ -150,6 +150,23 @@ enum Cmd {
         #[clap(long)]
         base: String,
     },
+
+    /// Shows the patch directory for a given product version
+    PatchDir {
+        #[clap(flatten)]
+        pv: ProductVersion,
+    },
+
+    /// Shows the worktree directory for a given product version
+    ///
+    /// This is the same value as `cargo patchable checkout` returns, but does not perform a checkout.
+    WorktreeDir {
+        #[clap(flatten)]
+        pv: ProductVersion,
+    },
+
+    /// Shwos the images repository root
+    ImagesDir,
 }
 
 #[derive(Debug, Snafu)]
@@ -181,7 +198,7 @@ pub enum Error {
     },
 
     #[snafu(display("failed to find images repository"))]
-    FindImagesRepo { source: git2::Error },
+    FindImagesRepo { source: repo::Error },
     #[snafu(display("images repository has no work directory"))]
     NoImagesRepoWorkdir,
 
@@ -246,7 +263,7 @@ fn main() -> Result<()> {
     .context(ConfigureGitLoggingSnafu)?;
 
     let opts = <Opts as clap::Parser>::parse();
-    let images_repo = Repository::discover(".").context(FindImagesRepoSnafu)?;
+    let images_repo = repo::discover_images_repo(".").context(FindImagesRepoSnafu)?;
     let images_repo_root = images_repo.workdir().context(NoImagesRepoWorkdirSnafu)?;
     match opts.cmd {
         Cmd::Checkout { pv, base_only } => {
@@ -415,6 +432,24 @@ fn main() -> Result<()> {
                 "created configuration for product version"
             );
         }
+
+        Cmd::PatchDir { pv } => {
+            let ctx = ProductVersionContext {
+                pv,
+                images_repo_root,
+            };
+            println!("{}", ctx.patch_dir().display());
+        }
+
+        Cmd::WorktreeDir { pv } => {
+            let ctx = ProductVersionContext {
+                pv,
+                images_repo_root,
+            };
+            println!("{}", ctx.worktree_root().display());
+        }
+
+        Cmd::ImagesDir => println!("{}", images_repo_root.display()),
     }
 
     Ok(())